neutrino-release: fix copy audio_dts.elf to release (Thanks BPanther for the hint)
- will only be copied if available

+++ make/neutrino-release.mk
@@ -128,7 +128,7 @@ neutrino-release-ufs912:
 	cp $(TARGET_DIR)/lib/modules/$(KERNEL_VER)/extra/stgfb/stmfb/stmcore-display-sti7111.ko $(RELEASE_DIR)/lib/modules/
 	cp $(SKEL_ROOT)/boot/video_7111.elf $(RELEASE_DIR)/lib/firmware/video.elf
 	cp $(SKEL_ROOT)/boot/audio_7111.elf $(RELEASE_DIR)/lib/firmware/audio.elf
-	cp $(SKEL_ROOT)/boot/audio_dts.elf $(RELEASE_DIR)/lib/firmware/audio_dts.elf
+	[ -e $(SKEL_ROOT)/boot/audio_dts.elf ] && cp $(SKEL_ROOT)/boot/audio_dts.elf $(RELEASE_DIR)/lib/firmware/audio_dts.elf || true
 	cp $(SKEL_ROOT)/firmware/component_7111_mb618.fw $(RELEASE_DIR)/lib/firmware/component.fw
 
 #
@@ -141,7 +141,7 @@ neutrino-release-ufs913:
 	cp $(TARGET_DIR)/lib/modules/$(KERNEL_VER)/extra/stgfb/stmfb/stmcore-display-sti7105.ko $(RELEASE_DIR)/lib/modules/
 	cp $(SKEL_ROOT)/boot/video_7105.elf $(RELEASE_DIR)/lib/firmware/video.elf
 	cp $(SKEL_ROOT)/boot/audio_7105.elf $(RELEASE_DIR)/lib/firmware/audio.elf
-	cp $(SKEL_ROOT)/boot/audio_dts.elf $(RELEASE_DIR)/lib/firmware/audio_dts.elf
+	[ -e $(SKEL_ROOT)/boot/audio_dts.elf ] && cp $(SKEL_ROOT)/boot/audio_dts.elf $(RELEASE_DIR)/lib/firmware/audio_dts.elf || true
 	cp $(SKEL_ROOT)/firmware/component_7105_pdk7105.fw $(RELEASE_DIR)/lib/firmware/component.fw
 	cp $(SKEL_ROOT)/firmware/dvb-fe-avl6222.fw $(RELEASE_DIR)/lib/firmware/
